Refactoring to extract base class.
