+++ src/clients/automower/automowerEventStreamClient.ts
@@ -2,7 +2,7 @@ import { PlatformLogger } from '../../diagnostics/platformLogger';
 import { AccessToken } from '../../model';
 import { WebSocketWrapper, WebSocketWrapperImpl } from '../../primitives/webSocketWrapper';
 import { PLUGIN_ID } from '../../settings';
-import { EventStreamClient } from '../eventStreamClient';
+import { AbstractEventStreamClient, EventStreamClient } from '../eventStreamClient';
 import { Battery, Calendar, Headlight, MowerMetadata, MowerState, Planner, Position } from './automowerClient';
 
 /**
@@ -85,44 +85,21 @@ export interface AutomowerEventStreamClient extends EventStreamClient {
     on(callback: (event: AutomowerEvent) => Promise<void>): void;
 }
 
-export class AutomowerEventStreamClientImpl implements AutomowerEventStreamClient {
-    private socket?: WebSocketWrapper;
+export class AutomowerEventStreamClientImpl extends AbstractEventStreamClient implements AutomowerEventStreamClient {
     private onMessageReceivedCallback?: (payload: AutomowerEvent) => void;
-
-    private onErrorReceivedCallback?: () => void;
-    private onConnectedCallback?: () => void;
-    private onDisconnectedCallback?: () => void;
-
-    private connecting = false;
-    private connected = false;
     private connectionId?: string;
     
-    public constructor(private baseUrl: string, private log: PlatformLogger) { }
-    
-    public open(token: AccessToken): void {
-        if (this.socket !== undefined) {
-            this.socket.close();
-        }
-
-        this.onConnecting();        
-        this.socket = this.createSocket(token);
-        
-        this.socket.on('message', this.onSocketMessageReceived.bind(this));
-        this.socket.on('error', this.onErrorReceived.bind(this));    
-        this.socket.on('close', this.onCloseReceived.bind(this));
+    public constructor(private baseUrl: string, log: PlatformLogger) { 
+        super(log);
     }
 
-    protected onConnecting(): void {
-        this.setConnecting(true);
-    }
-
-    protected createSocket(token: AccessToken): WebSocketWrapper {
-        return new WebSocketWrapperImpl(this.baseUrl, {
+    protected createSocket(token: AccessToken): Promise<WebSocketWrapper> {
+        return Promise.resolve(new WebSocketWrapperImpl(this.baseUrl, {
             headers: {
                 'Authorization': `Bearer ${token.value}`,
                 'X-Application-Id': PLUGIN_ID
             }
-        });
+        }));
     }
 
     public getConnectionId(): string | undefined {
@@ -131,46 +108,17 @@ export class AutomowerEventStreamClientImpl implements AutomowerEventStreamClien
 
     protected setConnectionId(value: string | undefined) {
         this.connectionId = value;
-    }
-
-    public ping(): void {
-        this.socket?.ping('ping');
-    }
-
-    public isConnected(): boolean {
-        return this.connected;
-    }
-
-    protected setConnected(value: boolean): void {
-        this.connected = value;
-    }
-
-    public isConnecting(): boolean {
-        return this.connecting;
-    }
-
-    protected setConnecting(value: boolean): void {
-        this.connecting = value;
-    }
+    }    
 
     public isCallbackSet(): boolean {
         return this.onMessageReceivedCallback !== undefined;
     }
 
-    protected onCloseReceived() {
-        if (this.isConnected()) {
-            this.setConnected(false);
-
-            if (this.onDisconnectedCallback !== undefined) {
-                try {
-                    this.onDisconnectedCallback();                
-                } catch (e) {
-                    this.log.error('ERROR_HANDLING_DISCONNECTED_EVENT', e);
-                }
-            }
-        } else if (this.connecting) {
-            this.setConnecting(false);
-        }        
+    protected override attachTo(socket: WebSocketWrapper): void {
+        super.attachTo(socket);
+        
+        socket.on('message', this.onSocketMessageReceived.bind(this));
+        socket.on('error', this.onErrorReceived.bind(this));
     }
     
     protected onErrorReceived(err: ErrorEvent): void {
@@ -180,13 +128,7 @@ export class AutomowerEventStreamClientImpl implements AutomowerEventStreamClien
             type: err.type
         });
 
-        if (this.onErrorReceivedCallback !== undefined) {
-            try {
-                this.onErrorReceivedCallback();
-            } catch (e) {
-                this.log.error('ERROR_HANDLING_ERROR_EVENT', e);
-            }
-        }
+        this.notifyErrorReceived();
     }
 
     protected onSocketMessageReceived(buffer: Buffer): void {
@@ -204,7 +146,7 @@ export class AutomowerEventStreamClientImpl implements AutomowerEventStreamClien
             } else {
                 const mowerEvent = data as AutomowerEvent;
                 if (mowerEvent.type !== undefined) {
-                    this.onMessageReceived(mowerEvent);
+                    this.notifyMessageReceived(mowerEvent);
                 }
             }
         } catch (e) {
@@ -214,47 +156,16 @@ export class AutomowerEventStreamClientImpl implements AutomowerEventStreamClien
 
     protected onConnectedReceived(event: ConnectedEvent): void {
         this.setConnectionId(event.connectionId);
-        this.setConnecting(false);
-        this.setConnected(true);
-        
-        if (this.onConnectedCallback !== undefined) {
-            try {
-                this.onConnectedCallback();
-            } catch (e) {
-                this.log.error('ERROR_HANDLING_CONNECTED_EVENT', e);
-            }
-        }
+        this.onConnectionSucceeded();
     }
 
-    protected onMessageReceived(event: AutomowerEvent): void {
-        if (this.onMessageReceivedCallback === undefined) {
-            return;
+    protected notifyMessageReceived(event: AutomowerEvent): void {
+        if (this.onMessageReceivedCallback !== undefined) {
+            this.onMessageReceivedCallback(event);
         }
-
-        this.onMessageReceivedCallback(event);
-    }
-
-    public close(): void {
-        if (this.socket === undefined) {
-            return;
-        }
-
-        this.socket.terminate();
     }
 
     public on(callback: (event: AutomowerEvent) => Promise<void>): void {        
         this.onMessageReceivedCallback = callback;
-    }
-
-    public onConnected(callback: () => Promise<void>): void {
-        this.onConnectedCallback = callback;
-    }
-
-    public onDisconnected(callback: () => Promise<void>): void {
-        this.onDisconnectedCallback = callback;
-    }
-
-    public onError(callback: () => Promise<void>): void {
-        this.onErrorReceivedCallback = callback;
     }
 }
